# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: COMPAON API
  version: 1.0.0
  contact:
    email: compaon@dalavina.com
    name: Compaon
    url: 'https://dalavina.com/compaon'
  license:
    name: EULA
    url: 'https://en.wikipedia.org/wiki/End-user_license_agreement'
  description: API definition for COMPAON
servers:
  - url: 'http://52.74.156.150:3000/api/v1'
paths:
...
  '/exhibitions/{exhibition_id}/booths/{booth_id}':
    put:
      security:
      - bearerAuth: []
      tags:
        - Exhibition
      summary: Update an existing booth
      description: Update an existing booth by Id
      operationId: updateBooth
      parameters:
        - name: exhibition_id
          description: ID to get exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: booth_id
          description: ID to get booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        description: Update an existing booth
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booth'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booth'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not found
                    details: []
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    get:
      security:
      - bearerAuth: []
      summary: Get booth by ID
      description: Return a single booth
      operationId: getBoothById
      parameters:
        - name: exhibition_id
          description: ID to get exhibition
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: booth_id
          description: ID to get booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
              example: 'booth_data.position_booth'
          description: |
            List of populable columns:
            - booth_images.booth_template_position
            - booth_projects.booth_template_position
            - booth_products.booth_template_position
            - booth_videos.booth_template_position
            - booth_template
            - location
            - location.space_template_location
        - in: query
          name: withDeleted
          schema:
            type: string
            example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booth'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
      tags:
        - Exhibition
...
  '/booth-organizations/{id}':
     get:
      security:
      - bearerAuth: []
      summary: Get booth organization by ID
      description: Return a single booth organization
      operationId: getBoothOrganizationById
      parameters:
      - name: id
        description: ID to get booth organization
        in: path
        required: true
        schema:
          type: string
          pattern: '^[0-9]+$'
      - in: query
        name: populate
        schema:
          type: array
          items:
            type: string
            example: 'booth_organization_images.image'
        description: |
          List of populatable columns:
            - booth_organization_template
            - booth_organization_template.booth_organization_template_positions
            - booth_organization_products.product
            - booth_organization_projects.project
            - booth_organization_videos.video
            - booth_organization_images.image
            - booth_organization_products.booth_organization_template_position
            - booth_organization_projects.booth_organization_template_position
            - booth_organization_videos.booth_organization_template_position
            - booth_organization_images.booth_organization_template_position
      - in: query
        name: withDeleted
        schema:
          type: string
          example: 'true'
      tags:
        - Booth Organization
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoothOrganization'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
     put:
      security:
      - bearerAuth: []
      tags:
        - Booth Organization
      summary: Update an existing booth organization
      description: Update an existing booth organization by Id
      operationId: updateBoothOrganization
      parameters:
        - name: id
          description: ID to update booth organization
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        description: Update an existing booth organization
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoothOrganization'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoothOrganization'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not found
                    details: []
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
...
  '/booths/{id}':
    get:
      security:
      - bearerAuth: []
      tags:
        - Booth
      summary: Get an existing booth
      description: Get an existing booth by Id
      operationId: getBooth
      parameters:
        - name: id
          description: ID to Get booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
              example: 'booth'
          description: |
            List of populatable columns:
              - booth_images.booth_template_position
              - booth_images.image
              - booth_projects.booth_template_position
              - booth_projects.project
              - booth_products.booth_template_position
              - booth_products.product
              - booth_videos.booth_template_position
              - booth_videos.video
              - booth_template
              - location
              - location.space_template_location
              - exhibition
              - meetings
        - in: query
          name: withDeleted
          schema:
            type: string
            example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booth'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    delete:
      security:
      - bearerAuth: []
      tags:
        - Booth
      summary: Delete an existing booth
      description: Delete an existing booth by Id
      operationId: deleteBooth
      parameters:
        - name: id
          description: ID to delete booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
...
components:
  schemas:
...
    BoothOrganization:
      type: object
      required:
        - id
        - booth_organization_template_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        position_x:
          type: number
          example: 1
        position_y:
          type: number
          example: 1
        position_z:
          type: number
          example: 1
        rotation_x:
          type: number
          example: 1
        rotation_y:
          type: number
          example: 1
        rotation_z:
          type: number
          example: 1
        booth_organization_template_id:
          type: integer
          writeOnly: true
          example: 1
        booth_organization_images:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/BoothOrganizationImage'
        booth_organization_videos:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/BoothOrganizationVideo'
        booth_organization_projects:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/BoothOrganizationProject'
        booth_organization_products:
          writeOnly: true
          type: array
          items:
            $ref: '#/components/schemas/BoothOrganizationProduct'
    BoothOrganizationImage:
      type: object
      required:
        - id
        - image_id
        - booth_organization_template_position_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        selected_media_id:
          type: integer
          writeOnly: true
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          example: 'base64 string'
        booth_organization_template_position_id:
          type: integer
          writeOnly: true
          example: 1
    BoothOrganizationVideo:
      type: object
      required:
        - id
        - video_id
        - booth_organization_template_position_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        selected_media_id:
          type: integer
          writeOnly: true
          example: 1
        video_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          example: 'base64 string'
        booth_organization_template_position_id:
          type: integer
          writeOnly: true
          example: 1
    BoothOrganizationProject:
      type: object
      required:
        - id
        - image_id
        - title
        - description
        - booth_organization_template_position_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        selected_media_id:
          type: integer
          writeOnly: true
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          example: 'base64 string'
        title:
          type: string
          maxLength: 255
          example: "Tên dự án"
        description:
          type: string
          maxLength: 20000
          example: "Mô tả dự án"
        booth_organization_template_position_id:
          type: integer
          writeOnly: true
          example: 1
    BoothOrganizationProduct:
      type: object
      required:
        - id
        - name
        - image_id
        - price
        - purchase_link
        - description
        - booth_organization_template_position_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 255
          example: "name"
        selected_media_id:
          type: integer
          writeOnly: true
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          example: 'base64 string'
        price:
          type: number
          format: float
          example: 20000
        purchase_link:
          type: string
          maxLength: 255
          example: 'https://link_mua_san_pham/ma_ha_noi.png'
        description:
          type: string
          maxLength: 500
          example: 'Mô tả sản phẩm'
        booth_organization_template_position_id:
          type: integer
          writeOnly: true
          example: 1
...
    ErrorResponseBody:
      type: object
      required:
        - message
        - details
      properties:
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
...
    Booth:
      type: object
      required:
        - id
        - name
        - enterprise_id
        - created_by
        - location_id
        - booth_template_id
      properties:
        id:
          type: integer
          example: 1
          readOnly: true
        name:
          type: string
        enterprise_id:
          type: integer
          example: 1
        created_by:
          type: integer
          readOnly: true
          example: 1
        created_date:
          type: string
          readOnly: true
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-01-01T00:00:00.000Z'
        location_id:
          type: integer
          writeOnly: true
          example: 1
        booth_template_id:
          type: integer
          writeOnly: true
          example: 1
        live_streams:
          type: array
          items:
            $ref: '#/components/schemas/LiveStream'
        booth_images:
          type: array
          items:
            $ref: '#/components/schemas/BoothImage'
        booth_videos:
          type: array
          items:
            $ref: '#/components/schemas/BoothVideo'
        booth_projects:
          type: array
          items:
            $ref: '#/components/schemas/BoothProject'
        booth_products:
          type: array
          items:
            $ref: '#/components/schemas/BoothProduct'
...
    LiveStream:
      type: object
      required:
        - id
        - link
        - time
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        link:
          type: string
          minLength: 1
          example: 'www.goole.com'
        time:
          type: string
          minLength: 1
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-12-01T00:00:00.000Z'
    BoothImage:
      type: object
      required:
        - id
        - booth_template_position_id
        - image_id
      properties:
        id:
          type: integer
          readOnly: true
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if image selected from data center
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if image selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          example: 1
    BoothVideo:
      type: object
      required:
        - id
        - booth_template_position_id
        - video_id
      properties:
        id:
          type: integer
          readOnly: true
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if video selected from data center
          example: 1
        video_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if video selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          example: 1
    BoothProject:
      type: object
      required:
        - id
        - booth_template_position_id
        - image_id
        - title
        - description
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        view:
          type: integer
          readOnly: true
          example: 1
        like:
          type: integer
          readOnly: true
          example: 1
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if video or image selected from data center
        image_id:
          type: integer
          readOnly: true
        media_data:
          type: string
          writeOnly: true
          description: This column exists if video or image selected from your local
          example: 'base64 string'
        title:
          type: string
          maxLength: 255
          example: 'Tên dự án'
        description:
          type: string
          maxLength: 20000
          example: 'Nội dung mô tả cho dự án'
        booth_template_position_id:
          type: integer
          writeOnly: true
          example: 1
...
    BoothProduct:
      type: object
      required:
        - id
        - name
        - price
        - purchase_link
        - booth_template_position_id
        - description
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 255
          example: 'mận Hà Nội'
        price:
          type: number
          format: float
          example: 35000
        purchase_link:
          type: string
          maxLength: 255
          example: 'https://link_mua_san_pham/ma_ha_noi.png'
        description:
          type: string
          maxLength: 500
          example: 'Mô tả sản phẩm'
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if image selected from data center
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if image selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          writeOnly: true
          example: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT